So i'm struggling on creating an instance of hand.
my idea is that an instance of Hand should instantiate by subtracting cards from an instance of Deck, then eventually ending it's lifecycle by adding itself to the instance discardpile. The discard pile will eventually cycle back to become a new instance of Deck.

In Hand, I started outlining a general way for this to happen...but it doesn't feel very right.

Would it be easier to make Hand not a class and make drawhand a function of Deck? To be clear - I'm confused how to do this either way.

In this commit, I also created a shipcard class to provide instances which work with instances of Hand to be the total stats of a player on any turn.

thanks in advance for the advice